Cache image URL when getting its details
This should reduce a lot of the overhead of calling `getImgDetails()`. It does appear to introduce an issue where we might accidentally cache a reference to the URL after it's been updated, but if changes are made through the `imgix.URL` API, the changes will be handled correctly, and if changes are made just by changing the `<img>`'s `src` attribute, imgix.js isn't gonna behave correctly anyway. Ergo: no regression here.