Added fast installation instructions
Although it might seem obvious, the `name` of the package is not always obvious. Most of times it is different, and some are the same as the repo name. I thought it was `imgix`, and unfortunately `npm` found a package named that way and installed. Took me 30 minutes to figure out...