Update polyfills.js
I found this patch useful to deal with prototypeJs, that library updates the native filter method to use the findAll prototypeJs method, which is not correct, so in order to prevent to this plugins to fail in some circunstancies I found this useful. 

Please review and let me know if we can merge into the main branch. Otherwise we might diverge from the original repo and use our own fork. 

Cheers