chore: add package-lock.json and prettier (#131)
## Description

This PR checks in `package-lock.json`, which is recommended to be checked in by npm.

This PR also adds code formatting with Prettier. This is accomplished by adding a `npm run pretty` command which anyone can run to format the source code and tests.

It is also recommended that PrettierCI be added to this repo to enforce code formatting.