fix!: deprecate resizing the height param when maintaining aspect ratio
BREAKING CHANGE: removes the logic in ix-src responsible for resizing the h (height) parameter to maintain aspect ratio when building the srcset attribute. With the help of the imgix aspect ratio parameter, users can now achieve the same effect with the inclusion of the ar parameter. This should yield slightly better performance, code cleanliness, and bring the library more in line with the imgix API.
Please note, the ar parameter should also be used with fit=crop to take effect