fix(getSize): calc app size if parent elem has columns
In order to avoid always setting `size` to the same value as the view-width
,we need to check that the parent node is not itself divided into columns.
We can use the `getComputedStyle` to get the parent's CSS `columnCount` &
gridTemplateColumns values.

For flex-box, grid, and other layout types, these property will be
populated and describe the ideal number of columns into which the
content of the element will be flowed.

If the value is numeric, ie not the strings `auto`, `null`, etc, we can
use it to calculate the desired width:

(elementWidth / parentColumnCount) -> newWidth

Read more on column-count property here:
https://developer.mozilla.org/en-US/docs/Web/CSS/column-count