fix(imgixtag): fix issue where sizes was set to null & use ix-sizes
This commit fixes an issue where ImgixTag.sizes() would return null if
sizes had previously been left as null or undefined. This could happen
when stubbing an image attribute for tests or if a user accidentally
set the value to null.

The commit also changes data-sizes to ix-sizes. This is done to avoid
possible collisions with the lazysizes.js library.